ia64: Increase TCB alignment to 64
As of glibc 2.42, the size of `struct pthread' is 1824 bytes, which can
fit right into the 32-byte boundary, defined by the default
TCB_ALIGNMENT.

However the stack guard and the point guard must be inserted right after
the structure, into the padding space. When the alignment is 32, there
will be no extra space for the stack guard and the pointer guard, thus
these two will be inserted into existing fields.

This effectively corrupts the struct pthread, causing free() to access
inaccessible memory during the release of the thread.